# [UI Tests] - Remove checks that are no longer needed for tests using self hosted sites