--- a/WordPress/UITests/Tests/LoginTests.swift
+++ b/WordPress/UITests/Tests/LoginTests.swift
@@ -18,7 +18,7 @@ class LoginTests: XCTestCase {
             .selectContinue()
             .proceedWith(email: WPUITestCredentials.testWPcomUserEmail)
             .proceedWithValidPassword()
-                    .verifyEpilogueDisplays(
+            .verifyEpilogueDisplays(
                 username: WPUITestCredentials.testWPcomUsername,
                 siteUrl: WPUITestCredentials.testWPcomSitePrimaryAddress
             )