@jostnes 2023-07-19
👋 @mokagio 

> Are those methods used anywhere else? Have you considered removing them altogether?

Yup, they are still being used by tests that are not using self-hosted sites. Though we are in the process of cleaning up the screens to use a standardized structure and in the process will be removing unused methods in an attempt to keep things cleaner and more maintainable. You'll start seeing those PRs soon.